% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscUtils-package.R
\docType{package}
\name{mscUtils-package}
\alias{mscUtils}
\alias{mscUtils-package}
\title{mscUtils: Utilities for MSc ecology literature reviews}
\description{
Tools for extracting information from published literature.
}
\author{
\strong{Maintainer}: Julian Flowers \email{jrf128@student.aru.ac.uk} (\href{https://orcid.org/0000-0001-5262-4526}{ORCID})

}
\keyword{internal}
